+++ sc/Program.cs
@@ -1,12 +1,47 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Santuryu.CodeAnalysis;
+using Santuryu.CodeAnalysis.Symbols;
+using Santuryu.CodeAnalysis.Syntax;
+using Santuryu.IO;
 
-namespace sc
+namespace Santuryu
 {
-    class Program
+    internal static class Program
     {
-        static void Main(string[] args)
+        private static void Main(string[] args)
         {
-            Console.WriteLine("Hello World!");
+            if (args.Length == 0)
+            {
+                Console.Error.WriteLine("usage: mc <source-paths>");
+                return;
+            }
+
+            if (args.Length > 1)
+            {
+                Console.WriteLine("error: only one path supported right now");
+                return;
+            }
+
+            var path = args.Single();
+
+            var text = File.ReadAllText(path);
+            var syntaxTree = SyntaxTree.Parse(text);
+
+            var compilation = new Compilation(syntaxTree);
+            var result = compilation.Evaluate(new Dictionary<VariableSymbol, object>());
+
+            if (!result.Diagnostics.Any())
+            {
+                if (result.Value != null)
+                    Console.WriteLine(result.Value);
+            }
+            else
+            {
+                Console.Error.WriteDiagnostics(result.Diagnostics, syntaxTree);
+            }
         }
     }
 }
