+++ sc/sc.cmd
@@ -1,0 +1,3 @@
+@echo off
+
+dotnet run --project .\src\sc\sc.csproj -- "%*"
