+++ sc/sc.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+dotnet run --project ./src/sc/sc.cspoj -- "$@"
